Add files via upload
We present our Buffer 5.0 project under the category College applications – ‘CUMMINS TASK SCHEDULER’. 
We have used array list implemented stacks to run our task scheduler. So first we ask the user to enter the number of tasks he needs to get done that day, as he enters the tasks it is displayed to the user so that he knows and then comes the operation menu asking the user 2 choices- whether he has completed the task and what his upcoming task is. So accordingly after finishing his task A, he marks it complete and then chooses option 2 to check what his next work is. After he is done with all the tasks the application is complete and will be run over for the next time.
The basic concept behind our project is that we initialize a stack, in which all the elements entered by the user are getting stored. then we declare a second stack pop the items in the original stack, and push them in the second one to organize them in the same order as the user entered since the stack uses the LIFO principle.